requirements.txt
shinylive
ipyleaflet
shinyswatch
micropip